Colorsensor Lázaro e Nicolas G1/colorsensor Lázaro e Nicolas/banco flask/requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy>=2
paho-mqtt
